Add a cd to LS_HOME to the init template.
Logstash's embedded elastic search looks for it's data directory under the current working directory when it is started.  Without a cd, this would be the root directory, and the server would fail to start when it could not create /data/elasticsearch.   Note, since they share the init template this change will also affect agent initialization, but that shouldn't cause problems.