improved vagrant support with v2 config options
remove manditory usage of rbenv for kibana, can toggle it with true/false.  defaults to true to retain author's initial intent.
useful for vagrant ( no need to sit around waiting for ruby to compile )

added parameters to set uid/gid for logstash agent/server for cases where a specific user is needed to access files ( say /var/log/messages on rhel is root:root ).  wanted it seperate to the user logstash is installed as to allow for some flexibility.  defaults are
to use the user/group parameters from base logstash to retain original intent.