Recipe to install the aws-plugin. Refs: #89
Downloads the version of the aws-plugin specified in attributes/cloud-aws.rb
Creates an elasticsearch.yml config file to join a ES cluster.
Appends aws-plugin to a Java Classpath for runit and init.

Source recipe now creates a flatjar by default.
Removes the first slash in sv-logstash_server_run which created an output of
two slashes (//)